Move IConsensusManager to own file
